Add article downloader and auxiliary classes
This commit introduces an `ArticleDownloader` class for downloading articles in chunks. It also includes the auxiliary classes and methods needed for this process such as `ArticleBlobFile` and `CustomNntpArticleResponse`. Changes in types and names, like `BlogOffset` from int to long in `ArticleHeader` and renaming DbSet property from `Article` to `ArticleHeader` in `GroupDbContext`, were also made correspondingly.

+++ src/Unflow/ArticleHeader.cs
@@ -22,6 +22,6 @@ public class ArticleHeader
     public DateTimeOffset? BodyDownloadedAt { get; set; }
 
     public int? BlobId { get; set; }
-    public int? BlogOffset { get; set; }
+    public long? BlogOffset { get; set; }
     public Guid? ParentId { get; set; }
 }
